[checkpoint_engine] move backend implementations to their backend packages
Per review: the SGLang in-process delta loader is the sglang rollout backend's
implementation of the sparse consumer, and the shard-view/diff/gather helpers
are the FSDP trainer backend's implementation of the shard export -- neither
belongs under checkpoint_engine/.

- delta_sync/sglang_loader.py -> workers/rollout/sglang_rollout/delta_loader.py
  (LOADER_FQN updated accordingly)
- delta_sync/sharded.py -> workers/engine/fsdp/sharded_delta.py
